requirements.txt.txt
Flask
requests
opencv-python
Pillow
